requirements.txt
# Orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# Data collection
yfinance>=0.2.40
fredapi>=0.5.1
requests>=2.31.0

# Feature engineering
pandas>=2.2.0
numpy>=1.26.0
pandas-ta>=0.3.14b

# ML - unsupervised
scikit-learn>=1.5.0
hdbscan>=0.8.33

# ML - supervised
xgboost>=2.0.0
lightgbm>=4.3.0

# Portfolio optimization
PyPortfolioOpt>=1.5.5
scipy>=1.13.0

# LLM — instalar solo el/los providers que uses
anthropic>=0.34.0         # Claude (Anthropic)
openai>=1.35.0            # GPT (OpenAI)
google-genai>=1.0.0       # Gemini (Google)

# Utils
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
rich>=13.7.0
typer>=0.12.0

# Bot / API (Render deploy)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-telegram-bot>=21.0
